Python_code/requirements.txt
future~=1.0.0
numpy~=1.26.4
opencv-python~=4.10.0.84
tensorflow~=2.9.3
tensorflow-gpu~=2.10.0
matplotlib~=3.7.1
pillow~=10.4.0
protobuf-5.29.4